cambios proyectos

+++ tetrace/models/tetrace_imputacion_proyectos.py
@@ -17,7 +17,7 @@ class imputacionhoras(models.Model):
     concepto_id = fields.Many2one('tetrace.imputacion_proyecto_nombre', string="Concepto")
     unidades = fields.Float(string="Unidades")
     precio = fields.Float(string="importe")
-    total = fields.Integer(string="total",compute="_compute_total_linea")
+    total = fields.Float(string="total",compute="_compute_total_linea")
     
     @api.depends('unidades','precio')
     def _compute_total_linea(self):
